component externalized 😎
